refactor: move models to the database dir
